% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_assertion_funs.R,
%   R/generated_base_report_funs.R, R/generated_report_fun_variants.R,
%   R/generated_test_funs.R, R/report_funs.R
\name{assert_atom_is_in_set}
\alias{assert_atom_is_in_set}
\alias{assert_dir_exists}
\alias{assert_file_exists}
\alias{assert_has_class}
\alias{assert_has_length}
\alias{assert_has_names}
\alias{assert_has_one_of_classes}
\alias{assert_has_only_names}
\alias{assert_inherits}
\alias{assert_is_atom}
\alias{assert_is_between_exclusive}
\alias{assert_is_between_inclusive}
\alias{assert_is_call}
\alias{assert_is_character}
\alias{assert_is_character_atom}
\alias{assert_is_character_matrix}
\alias{assert_is_character_nonNA_atom}
\alias{assert_is_character_nonNA_matrix}
\alias{assert_is_character_nonNA_vector}
\alias{assert_is_character_vector}
\alias{assert_is_data.frame}
\alias{assert_is_data.frame_with_required_names}
\alias{assert_is_data.table}
\alias{assert_is_data.table_with_required_names}
\alias{assert_is_data_table}
\alias{assert_is_data_table_with_required_names}
\alias{assert_is_Date}
\alias{assert_is_Date_atom}
\alias{assert_is_Date_matrix}
\alias{assert_is_Date_nonNA_atom}
\alias{assert_is_Date_nonNA_matrix}
\alias{assert_is_Date_nonNA_vector}
\alias{assert_is_Date_vector}
\alias{assert_is_double}
\alias{assert_is_double_atom}
\alias{assert_is_double_gtezero_atom}
\alias{assert_is_double_gtezero_matrix}
\alias{assert_is_double_gtezero_vector}
\alias{assert_is_double_gtzero_atom}
\alias{assert_is_double_gtzero_matrix}
\alias{assert_is_double_gtzero_vector}
\alias{assert_is_double_ltezero_atom}
\alias{assert_is_double_ltezero_matrix}
\alias{assert_is_double_ltezero_vector}
\alias{assert_is_double_ltzero_atom}
\alias{assert_is_double_ltzero_matrix}
\alias{assert_is_double_ltzero_vector}
\alias{assert_is_double_matrix}
\alias{assert_is_double_nonNA_atom}
\alias{assert_is_double_nonNA_gtezero_atom}
\alias{assert_is_double_nonNA_gtezero_matrix}
\alias{assert_is_double_nonNA_gtezero_vector}
\alias{assert_is_double_nonNA_gtzero_atom}
\alias{assert_is_double_nonNA_gtzero_matrix}
\alias{assert_is_double_nonNA_gtzero_vector}
\alias{assert_is_double_nonNA_ltezero_atom}
\alias{assert_is_double_nonNA_ltezero_matrix}
\alias{assert_is_double_nonNA_ltezero_vector}
\alias{assert_is_double_nonNA_ltzero_atom}
\alias{assert_is_double_nonNA_ltzero_matrix}
\alias{assert_is_double_nonNA_ltzero_vector}
\alias{assert_is_double_nonNA_matrix}
\alias{assert_is_double_nonNA_vector}
\alias{assert_is_double_vector}
\alias{assert_is_environment}
\alias{assert_is_expression}
\alias{assert_is_factor}
\alias{assert_is_factor_atom}
\alias{assert_is_factor_matrix}
\alias{assert_is_factor_nonNA_atom}
\alias{assert_is_factor_nonNA_matrix}
\alias{assert_is_factor_nonNA_vector}
\alias{assert_is_factor_vector}
\alias{assert_is_factor_with_levels}
\alias{assert_is_function}
\alias{assert_is_gt}
\alias{assert_is_gte}
\alias{assert_is_gtezero}
\alias{assert_is_gtzero}
\alias{assert_is_integer}
\alias{assert_is_integer_atom}
\alias{assert_is_integer_gtezero_atom}
\alias{assert_is_integer_gtezero_matrix}
\alias{assert_is_integer_gtezero_vector}
\alias{assert_is_integer_gtzero_atom}
\alias{assert_is_integer_gtzero_matrix}
\alias{assert_is_integer_gtzero_vector}
\alias{assert_is_integer_ltezero_atom}
\alias{assert_is_integer_ltezero_matrix}
\alias{assert_is_integer_ltezero_vector}
\alias{assert_is_integer_ltzero_atom}
\alias{assert_is_integer_ltzero_matrix}
\alias{assert_is_integer_ltzero_vector}
\alias{assert_is_integer_matrix}
\alias{assert_is_integer_nonNA_atom}
\alias{assert_is_integer_nonNA_gtezero_atom}
\alias{assert_is_integer_nonNA_gtezero_matrix}
\alias{assert_is_integer_nonNA_gtezero_vector}
\alias{assert_is_integer_nonNA_gtzero_atom}
\alias{assert_is_integer_nonNA_gtzero_matrix}
\alias{assert_is_integer_nonNA_gtzero_vector}
\alias{assert_is_integer_nonNA_ltezero_atom}
\alias{assert_is_integer_nonNA_ltezero_matrix}
\alias{assert_is_integer_nonNA_ltezero_vector}
\alias{assert_is_integer_nonNA_ltzero_atom}
\alias{assert_is_integer_nonNA_ltzero_matrix}
\alias{assert_is_integer_nonNA_ltzero_vector}
\alias{assert_is_integer_nonNA_matrix}
\alias{assert_is_integer_nonNA_vector}
\alias{assert_is_integer_vector}
\alias{assert_is_language_object}
\alias{assert_is_list}
\alias{assert_is_logical}
\alias{assert_is_logical_atom}
\alias{assert_is_logical_matrix}
\alias{assert_is_logical_nonNA_atom}
\alias{assert_is_logical_nonNA_matrix}
\alias{assert_is_logical_nonNA_vector}
\alias{assert_is_logical_vector}
\alias{assert_is_lt}
\alias{assert_is_lte}
\alias{assert_is_ltezero}
\alias{assert_is_ltzero}
\alias{assert_is_matrix}
\alias{assert_is_name}
\alias{assert_is_named}
\alias{assert_is_named_list}
\alias{assert_is_nonNA}
\alias{assert_is_NULL}
\alias{assert_is_number}
\alias{assert_is_number_atom}
\alias{assert_is_number_gtezero_atom}
\alias{assert_is_number_gtezero_matrix}
\alias{assert_is_number_gtezero_vector}
\alias{assert_is_number_gtzero_atom}
\alias{assert_is_number_gtzero_matrix}
\alias{assert_is_number_gtzero_vector}
\alias{assert_is_number_ltezero_atom}
\alias{assert_is_number_ltezero_matrix}
\alias{assert_is_number_ltezero_vector}
\alias{assert_is_number_ltzero_atom}
\alias{assert_is_number_ltzero_matrix}
\alias{assert_is_number_ltzero_vector}
\alias{assert_is_number_matrix}
\alias{assert_is_number_nonNA_atom}
\alias{assert_is_number_nonNA_gtezero_atom}
\alias{assert_is_number_nonNA_gtezero_matrix}
\alias{assert_is_number_nonNA_gtezero_vector}
\alias{assert_is_number_nonNA_gtzero_atom}
\alias{assert_is_number_nonNA_gtzero_matrix}
\alias{assert_is_number_nonNA_gtzero_vector}
\alias{assert_is_number_nonNA_ltezero_atom}
\alias{assert_is_number_nonNA_ltezero_matrix}
\alias{assert_is_number_nonNA_ltezero_vector}
\alias{assert_is_number_nonNA_ltzero_atom}
\alias{assert_is_number_nonNA_ltzero_matrix}
\alias{assert_is_number_nonNA_ltzero_vector}
\alias{assert_is_number_nonNA_matrix}
\alias{assert_is_number_nonNA_vector}
\alias{assert_is_number_vector}
\alias{assert_is_numeric}
\alias{assert_is_of_length}
\alias{assert_is_uniquely_named}
\alias{assert_is_uniquely_named_list}
\alias{assert_is_vector}
\alias{assert_vector_elems_are_in_set}
\alias{report_atom_is_in_set}
\alias{report_dir_exists}
\alias{report_file_exists}
\alias{report_has_class}
\alias{report_has_length}
\alias{report_has_names}
\alias{report_has_one_of_classes}
\alias{report_has_only_names}
\alias{report_inherits}
\alias{report_is_atom}
\alias{report_is_between_exclusive}
\alias{report_is_between_inclusive}
\alias{report_is_call}
\alias{report_is_character}
\alias{report_is_data.frame}
\alias{report_is_data.frame_with_required_names}
\alias{report_is_data.table}
\alias{report_is_data.table_with_required_names}
\alias{report_is_data_table}
\alias{report_is_data_table_with_required_names}
\alias{report_is_Date}
\alias{report_is_double}
\alias{report_is_environment}
\alias{report_is_expression}
\alias{report_is_factor}
\alias{report_is_factor_with_levels}
\alias{report_is_function}
\alias{report_is_gt}
\alias{report_is_gte}
\alias{report_is_gtezero}
\alias{report_is_gtzero}
\alias{report_is_integer}
\alias{report_is_language_object}
\alias{report_is_list}
\alias{report_is_logical}
\alias{report_is_lt}
\alias{report_is_lte}
\alias{report_is_ltezero}
\alias{report_is_ltzero}
\alias{report_is_matrix}
\alias{report_is_name}
\alias{report_is_named}
\alias{report_is_named_list}
\alias{report_is_nonNA}
\alias{report_is_NULL}
\alias{report_is_number}
\alias{report_is_numeric}
\alias{report_is_of_length}
\alias{report_is_uniquely_named}
\alias{report_is_uniquely_named_list}
\alias{report_is_vector}
\alias{report_vector_elems_are_in_set}
\alias{report_is_Date_atom}
\alias{report_is_Date_matrix}
\alias{report_is_Date_vector}
\alias{report_is_Date_nonNA_atom}
\alias{report_is_Date_nonNA_matrix}
\alias{report_is_Date_nonNA_vector}
\alias{report_is_character_atom}
\alias{report_is_character_matrix}
\alias{report_is_character_vector}
\alias{report_is_character_nonNA_atom}
\alias{report_is_character_nonNA_matrix}
\alias{report_is_character_nonNA_vector}
\alias{report_is_double_atom}
\alias{report_is_double_matrix}
\alias{report_is_double_vector}
\alias{report_is_double_gtezero_atom}
\alias{report_is_double_gtezero_matrix}
\alias{report_is_double_gtezero_vector}
\alias{report_is_double_gtzero_atom}
\alias{report_is_double_gtzero_matrix}
\alias{report_is_double_gtzero_vector}
\alias{report_is_double_ltezero_atom}
\alias{report_is_double_ltezero_matrix}
\alias{report_is_double_ltezero_vector}
\alias{report_is_double_ltzero_atom}
\alias{report_is_double_ltzero_matrix}
\alias{report_is_double_ltzero_vector}
\alias{report_is_double_nonNA_atom}
\alias{report_is_double_nonNA_matrix}
\alias{report_is_double_nonNA_vector}
\alias{report_is_double_nonNA_gtezero_atom}
\alias{report_is_double_nonNA_gtezero_matrix}
\alias{report_is_double_nonNA_gtezero_vector}
\alias{report_is_double_nonNA_gtzero_atom}
\alias{report_is_double_nonNA_gtzero_matrix}
\alias{report_is_double_nonNA_gtzero_vector}
\alias{report_is_double_nonNA_ltezero_atom}
\alias{report_is_double_nonNA_ltezero_matrix}
\alias{report_is_double_nonNA_ltezero_vector}
\alias{report_is_double_nonNA_ltzero_atom}
\alias{report_is_double_nonNA_ltzero_matrix}
\alias{report_is_double_nonNA_ltzero_vector}
\alias{report_is_factor_atom}
\alias{report_is_factor_matrix}
\alias{report_is_factor_vector}
\alias{report_is_factor_nonNA_atom}
\alias{report_is_factor_nonNA_matrix}
\alias{report_is_factor_nonNA_vector}
\alias{report_is_integer_atom}
\alias{report_is_integer_matrix}
\alias{report_is_integer_vector}
\alias{report_is_integer_gtezero_atom}
\alias{report_is_integer_gtezero_matrix}
\alias{report_is_integer_gtezero_vector}
\alias{report_is_integer_gtzero_atom}
\alias{report_is_integer_gtzero_matrix}
\alias{report_is_integer_gtzero_vector}
\alias{report_is_integer_ltezero_atom}
\alias{report_is_integer_ltezero_matrix}
\alias{report_is_integer_ltezero_vector}
\alias{report_is_integer_ltzero_atom}
\alias{report_is_integer_ltzero_matrix}
\alias{report_is_integer_ltzero_vector}
\alias{report_is_integer_nonNA_atom}
\alias{report_is_integer_nonNA_matrix}
\alias{report_is_integer_nonNA_vector}
\alias{report_is_integer_nonNA_gtezero_atom}
\alias{report_is_integer_nonNA_gtezero_matrix}
\alias{report_is_integer_nonNA_gtezero_vector}
\alias{report_is_integer_nonNA_gtzero_atom}
\alias{report_is_integer_nonNA_gtzero_matrix}
\alias{report_is_integer_nonNA_gtzero_vector}
\alias{report_is_integer_nonNA_ltezero_atom}
\alias{report_is_integer_nonNA_ltezero_matrix}
\alias{report_is_integer_nonNA_ltezero_vector}
\alias{report_is_integer_nonNA_ltzero_atom}
\alias{report_is_integer_nonNA_ltzero_matrix}
\alias{report_is_integer_nonNA_ltzero_vector}
\alias{report_is_logical_atom}
\alias{report_is_logical_matrix}
\alias{report_is_logical_vector}
\alias{report_is_logical_nonNA_atom}
\alias{report_is_logical_nonNA_matrix}
\alias{report_is_logical_nonNA_vector}
\alias{report_is_number_atom}
\alias{report_is_number_matrix}
\alias{report_is_number_vector}
\alias{report_is_number_gtezero_atom}
\alias{report_is_number_gtezero_matrix}
\alias{report_is_number_gtezero_vector}
\alias{report_is_number_gtzero_atom}
\alias{report_is_number_gtzero_matrix}
\alias{report_is_number_gtzero_vector}
\alias{report_is_number_ltezero_atom}
\alias{report_is_number_ltezero_matrix}
\alias{report_is_number_ltezero_vector}
\alias{report_is_number_ltzero_atom}
\alias{report_is_number_ltzero_matrix}
\alias{report_is_number_ltzero_vector}
\alias{report_is_number_nonNA_atom}
\alias{report_is_number_nonNA_matrix}
\alias{report_is_number_nonNA_vector}
\alias{report_is_number_nonNA_gtezero_atom}
\alias{report_is_number_nonNA_gtezero_matrix}
\alias{report_is_number_nonNA_gtezero_vector}
\alias{report_is_number_nonNA_gtzero_atom}
\alias{report_is_number_nonNA_gtzero_matrix}
\alias{report_is_number_nonNA_gtzero_vector}
\alias{report_is_number_nonNA_ltezero_atom}
\alias{report_is_number_nonNA_ltezero_matrix}
\alias{report_is_number_nonNA_ltezero_vector}
\alias{report_is_number_nonNA_ltzero_atom}
\alias{report_is_number_nonNA_ltzero_matrix}
\alias{report_is_number_nonNA_ltzero_vector}
\alias{test_atom_is_in_set}
\alias{test_dir_exists}
\alias{test_file_exists}
\alias{test_has_class}
\alias{test_has_length}
\alias{test_has_names}
\alias{test_has_one_of_classes}
\alias{test_has_only_names}
\alias{test_inherits}
\alias{test_is_atom}
\alias{test_is_between_exclusive}
\alias{test_is_between_inclusive}
\alias{test_is_call}
\alias{test_is_character}
\alias{test_is_character_atom}
\alias{test_is_character_matrix}
\alias{test_is_character_nonNA_atom}
\alias{test_is_character_nonNA_matrix}
\alias{test_is_character_nonNA_vector}
\alias{test_is_character_vector}
\alias{test_is_data.frame}
\alias{test_is_data.frame_with_required_names}
\alias{test_is_data.table}
\alias{test_is_data.table_with_required_names}
\alias{test_is_data_table}
\alias{test_is_data_table_with_required_names}
\alias{test_is_Date}
\alias{test_is_Date_atom}
\alias{test_is_Date_matrix}
\alias{test_is_Date_nonNA_atom}
\alias{test_is_Date_nonNA_matrix}
\alias{test_is_Date_nonNA_vector}
\alias{test_is_Date_vector}
\alias{test_is_double}
\alias{test_is_double_atom}
\alias{test_is_double_gtezero_atom}
\alias{test_is_double_gtezero_matrix}
\alias{test_is_double_gtezero_vector}
\alias{test_is_double_gtzero_atom}
\alias{test_is_double_gtzero_matrix}
\alias{test_is_double_gtzero_vector}
\alias{test_is_double_ltezero_atom}
\alias{test_is_double_ltezero_matrix}
\alias{test_is_double_ltezero_vector}
\alias{test_is_double_ltzero_atom}
\alias{test_is_double_ltzero_matrix}
\alias{test_is_double_ltzero_vector}
\alias{test_is_double_matrix}
\alias{test_is_double_nonNA_atom}
\alias{test_is_double_nonNA_gtezero_atom}
\alias{test_is_double_nonNA_gtezero_matrix}
\alias{test_is_double_nonNA_gtezero_vector}
\alias{test_is_double_nonNA_gtzero_atom}
\alias{test_is_double_nonNA_gtzero_matrix}
\alias{test_is_double_nonNA_gtzero_vector}
\alias{test_is_double_nonNA_ltezero_atom}
\alias{test_is_double_nonNA_ltezero_matrix}
\alias{test_is_double_nonNA_ltezero_vector}
\alias{test_is_double_nonNA_ltzero_atom}
\alias{test_is_double_nonNA_ltzero_matrix}
\alias{test_is_double_nonNA_ltzero_vector}
\alias{test_is_double_nonNA_matrix}
\alias{test_is_double_nonNA_vector}
\alias{test_is_double_vector}
\alias{test_is_environment}
\alias{test_is_expression}
\alias{test_is_factor}
\alias{test_is_factor_atom}
\alias{test_is_factor_matrix}
\alias{test_is_factor_nonNA_atom}
\alias{test_is_factor_nonNA_matrix}
\alias{test_is_factor_nonNA_vector}
\alias{test_is_factor_vector}
\alias{test_is_factor_with_levels}
\alias{test_is_function}
\alias{test_is_gt}
\alias{test_is_gte}
\alias{test_is_gtezero}
\alias{test_is_gtzero}
\alias{test_is_integer}
\alias{test_is_integer_atom}
\alias{test_is_integer_gtezero_atom}
\alias{test_is_integer_gtezero_matrix}
\alias{test_is_integer_gtezero_vector}
\alias{test_is_integer_gtzero_atom}
\alias{test_is_integer_gtzero_matrix}
\alias{test_is_integer_gtzero_vector}
\alias{test_is_integer_ltezero_atom}
\alias{test_is_integer_ltezero_matrix}
\alias{test_is_integer_ltezero_vector}
\alias{test_is_integer_ltzero_atom}
\alias{test_is_integer_ltzero_matrix}
\alias{test_is_integer_ltzero_vector}
\alias{test_is_integer_matrix}
\alias{test_is_integer_nonNA_atom}
\alias{test_is_integer_nonNA_gtezero_atom}
\alias{test_is_integer_nonNA_gtezero_matrix}
\alias{test_is_integer_nonNA_gtezero_vector}
\alias{test_is_integer_nonNA_gtzero_atom}
\alias{test_is_integer_nonNA_gtzero_matrix}
\alias{test_is_integer_nonNA_gtzero_vector}
\alias{test_is_integer_nonNA_ltezero_atom}
\alias{test_is_integer_nonNA_ltezero_matrix}
\alias{test_is_integer_nonNA_ltezero_vector}
\alias{test_is_integer_nonNA_ltzero_atom}
\alias{test_is_integer_nonNA_ltzero_matrix}
\alias{test_is_integer_nonNA_ltzero_vector}
\alias{test_is_integer_nonNA_matrix}
\alias{test_is_integer_nonNA_vector}
\alias{test_is_integer_vector}
\alias{test_is_language_object}
\alias{test_is_list}
\alias{test_is_logical}
\alias{test_is_logical_atom}
\alias{test_is_logical_matrix}
\alias{test_is_logical_nonNA_atom}
\alias{test_is_logical_nonNA_matrix}
\alias{test_is_logical_nonNA_vector}
\alias{test_is_logical_vector}
\alias{test_is_lt}
\alias{test_is_lte}
\alias{test_is_ltezero}
\alias{test_is_ltzero}
\alias{test_is_matrix}
\alias{test_is_name}
\alias{test_is_named}
\alias{test_is_named_list}
\alias{test_is_nonNA}
\alias{test_is_NULL}
\alias{test_is_number}
\alias{test_is_number_atom}
\alias{test_is_number_gtezero_atom}
\alias{test_is_number_gtezero_matrix}
\alias{test_is_number_gtezero_vector}
\alias{test_is_number_gtzero_atom}
\alias{test_is_number_gtzero_matrix}
\alias{test_is_number_gtzero_vector}
\alias{test_is_number_ltezero_atom}
\alias{test_is_number_ltezero_matrix}
\alias{test_is_number_ltezero_vector}
\alias{test_is_number_ltzero_atom}
\alias{test_is_number_ltzero_matrix}
\alias{test_is_number_ltzero_vector}
\alias{test_is_number_matrix}
\alias{test_is_number_nonNA_atom}
\alias{test_is_number_nonNA_gtezero_atom}
\alias{test_is_number_nonNA_gtezero_matrix}
\alias{test_is_number_nonNA_gtezero_vector}
\alias{test_is_number_nonNA_gtzero_atom}
\alias{test_is_number_nonNA_gtzero_matrix}
\alias{test_is_number_nonNA_gtzero_vector}
\alias{test_is_number_nonNA_ltezero_atom}
\alias{test_is_number_nonNA_ltezero_matrix}
\alias{test_is_number_nonNA_ltezero_vector}
\alias{test_is_number_nonNA_ltzero_atom}
\alias{test_is_number_nonNA_ltzero_matrix}
\alias{test_is_number_nonNA_ltzero_vector}
\alias{test_is_number_nonNA_matrix}
\alias{test_is_number_nonNA_vector}
\alias{test_is_number_vector}
\alias{test_is_numeric}
\alias{test_is_of_length}
\alias{test_is_uniquely_named}
\alias{test_is_uniquely_named_list}
\alias{test_is_vector}
\alias{test_vector_elems_are_in_set}
\alias{assertions}
\alias{assert_is_one_of}
\title{Assertions}
\usage{
assert_atom_is_in_set(x, x_nm = NULL, set)

assert_dir_exists(x, x_nm = NULL)

assert_file_exists(x, x_nm = NULL)

assert_has_class(x, x_nm = NULL, required_class)

assert_has_length(x, x_nm = NULL, expected_length)

assert_has_names(x, x_nm = NULL, required_names)

assert_has_one_of_classes(x, x_nm = NULL, classes)

assert_has_only_names(x, x_nm = NULL, required_names)

assert_inherits(x, x_nm = NULL, required_class)

assert_is_atom(x, x_nm = NULL)

assert_is_between_exclusive(x, x_nm = NULL, lo, hi)

assert_is_between_inclusive(x, x_nm = NULL, lo, hi)

assert_is_call(x, x_nm = NULL)

assert_is_character(x, x_nm = NULL)

assert_is_character_atom(x, x_nm = NULL)

assert_is_character_matrix(x, x_nm = NULL)

assert_is_character_nonNA_atom(x, x_nm = NULL)

assert_is_character_nonNA_matrix(x, x_nm = NULL)

assert_is_character_nonNA_vector(x, x_nm = NULL)

assert_is_character_vector(x, x_nm = NULL)

assert_is_data.frame(x, x_nm = NULL)

assert_is_data.frame_with_required_names(x, x_nm = NULL, required_names)

assert_is_data.table(x, x_nm = NULL)

assert_is_data.table_with_required_names(x, x_nm = NULL, required_names)

assert_is_data_table(x, x_nm = NULL)

assert_is_data_table_with_required_names(x, x_nm = NULL, required_names)

assert_is_Date(x, x_nm = NULL)

assert_is_Date_atom(x, x_nm = NULL)

assert_is_Date_matrix(x, x_nm = NULL)

assert_is_Date_nonNA_atom(x, x_nm = NULL)

assert_is_Date_nonNA_matrix(x, x_nm = NULL)

assert_is_Date_nonNA_vector(x, x_nm = NULL)

assert_is_Date_vector(x, x_nm = NULL)

assert_is_double(x, x_nm = NULL)

assert_is_double_atom(x, x_nm = NULL)

assert_is_double_gtezero_atom(x, x_nm = NULL)

assert_is_double_gtezero_matrix(x, x_nm = NULL)

assert_is_double_gtezero_vector(x, x_nm = NULL)

assert_is_double_gtzero_atom(x, x_nm = NULL)

assert_is_double_gtzero_matrix(x, x_nm = NULL)

assert_is_double_gtzero_vector(x, x_nm = NULL)

assert_is_double_ltezero_atom(x, x_nm = NULL)

assert_is_double_ltezero_matrix(x, x_nm = NULL)

assert_is_double_ltezero_vector(x, x_nm = NULL)

assert_is_double_ltzero_atom(x, x_nm = NULL)

assert_is_double_ltzero_matrix(x, x_nm = NULL)

assert_is_double_ltzero_vector(x, x_nm = NULL)

assert_is_double_matrix(x, x_nm = NULL)

assert_is_double_nonNA_atom(x, x_nm = NULL)

assert_is_double_nonNA_gtezero_atom(x, x_nm = NULL)

assert_is_double_nonNA_gtezero_matrix(x, x_nm = NULL)

assert_is_double_nonNA_gtezero_vector(x, x_nm = NULL)

assert_is_double_nonNA_gtzero_atom(x, x_nm = NULL)

assert_is_double_nonNA_gtzero_matrix(x, x_nm = NULL)

assert_is_double_nonNA_gtzero_vector(x, x_nm = NULL)

assert_is_double_nonNA_ltezero_atom(x, x_nm = NULL)

assert_is_double_nonNA_ltezero_matrix(x, x_nm = NULL)

assert_is_double_nonNA_ltezero_vector(x, x_nm = NULL)

assert_is_double_nonNA_ltzero_atom(x, x_nm = NULL)

assert_is_double_nonNA_ltzero_matrix(x, x_nm = NULL)

assert_is_double_nonNA_ltzero_vector(x, x_nm = NULL)

assert_is_double_nonNA_matrix(x, x_nm = NULL)

assert_is_double_nonNA_vector(x, x_nm = NULL)

assert_is_double_vector(x, x_nm = NULL)

assert_is_environment(x, x_nm = NULL)

assert_is_expression(x, x_nm = NULL)

assert_is_factor(x, x_nm = NULL)

assert_is_factor_atom(x, x_nm = NULL)

assert_is_factor_matrix(x, x_nm = NULL)

assert_is_factor_nonNA_atom(x, x_nm = NULL)

assert_is_factor_nonNA_matrix(x, x_nm = NULL)

assert_is_factor_nonNA_vector(x, x_nm = NULL)

assert_is_factor_vector(x, x_nm = NULL)

assert_is_factor_with_levels(x, x_nm = NULL, expected_levels)

assert_is_function(x, x_nm = NULL)

assert_is_gt(x, x_nm = NULL, lo)

assert_is_gte(x, x_nm = NULL, lo)

assert_is_gtezero(x, x_nm = NULL)

assert_is_gtzero(x, x_nm = NULL)

assert_is_integer(x, x_nm = NULL)

assert_is_integer_atom(x, x_nm = NULL)

assert_is_integer_gtezero_atom(x, x_nm = NULL)

assert_is_integer_gtezero_matrix(x, x_nm = NULL)

assert_is_integer_gtezero_vector(x, x_nm = NULL)

assert_is_integer_gtzero_atom(x, x_nm = NULL)

assert_is_integer_gtzero_matrix(x, x_nm = NULL)

assert_is_integer_gtzero_vector(x, x_nm = NULL)

assert_is_integer_ltezero_atom(x, x_nm = NULL)

assert_is_integer_ltezero_matrix(x, x_nm = NULL)

assert_is_integer_ltezero_vector(x, x_nm = NULL)

assert_is_integer_ltzero_atom(x, x_nm = NULL)

assert_is_integer_ltzero_matrix(x, x_nm = NULL)

assert_is_integer_ltzero_vector(x, x_nm = NULL)

assert_is_integer_matrix(x, x_nm = NULL)

assert_is_integer_nonNA_atom(x, x_nm = NULL)

assert_is_integer_nonNA_gtezero_atom(x, x_nm = NULL)

assert_is_integer_nonNA_gtezero_matrix(x, x_nm = NULL)

assert_is_integer_nonNA_gtezero_vector(x, x_nm = NULL)

assert_is_integer_nonNA_gtzero_atom(x, x_nm = NULL)

assert_is_integer_nonNA_gtzero_matrix(x, x_nm = NULL)

assert_is_integer_nonNA_gtzero_vector(x, x_nm = NULL)

assert_is_integer_nonNA_ltezero_atom(x, x_nm = NULL)

assert_is_integer_nonNA_ltezero_matrix(x, x_nm = NULL)

assert_is_integer_nonNA_ltezero_vector(x, x_nm = NULL)

assert_is_integer_nonNA_ltzero_atom(x, x_nm = NULL)

assert_is_integer_nonNA_ltzero_matrix(x, x_nm = NULL)

assert_is_integer_nonNA_ltzero_vector(x, x_nm = NULL)

assert_is_integer_nonNA_matrix(x, x_nm = NULL)

assert_is_integer_nonNA_vector(x, x_nm = NULL)

assert_is_integer_vector(x, x_nm = NULL)

assert_is_language_object(x, x_nm = NULL)

assert_is_list(x, x_nm = NULL)

assert_is_logical(x, x_nm = NULL)

assert_is_logical_atom(x, x_nm = NULL)

assert_is_logical_matrix(x, x_nm = NULL)

assert_is_logical_nonNA_atom(x, x_nm = NULL)

assert_is_logical_nonNA_matrix(x, x_nm = NULL)

assert_is_logical_nonNA_vector(x, x_nm = NULL)

assert_is_logical_vector(x, x_nm = NULL)

assert_is_lt(x, x_nm = NULL, lo, hi)

assert_is_lte(x, x_nm = NULL, hi)

assert_is_ltezero(x, x_nm = NULL)

assert_is_ltzero(x, x_nm = NULL)

assert_is_matrix(x, x_nm = NULL)

assert_is_name(x, x_nm = NULL)

assert_is_named(x, x_nm = NULL)

assert_is_named_list(x, x_nm = NULL)

assert_is_nonNA(x, x_nm = NULL)

assert_is_NULL(x, x_nm = NULL)

assert_is_number(x, x_nm = NULL)

assert_is_number_atom(x, x_nm = NULL)

assert_is_number_gtezero_atom(x, x_nm = NULL)

assert_is_number_gtezero_matrix(x, x_nm = NULL)

assert_is_number_gtezero_vector(x, x_nm = NULL)

assert_is_number_gtzero_atom(x, x_nm = NULL)

assert_is_number_gtzero_matrix(x, x_nm = NULL)

assert_is_number_gtzero_vector(x, x_nm = NULL)

assert_is_number_ltezero_atom(x, x_nm = NULL)

assert_is_number_ltezero_matrix(x, x_nm = NULL)

assert_is_number_ltezero_vector(x, x_nm = NULL)

assert_is_number_ltzero_atom(x, x_nm = NULL)

assert_is_number_ltzero_matrix(x, x_nm = NULL)

assert_is_number_ltzero_vector(x, x_nm = NULL)

assert_is_number_matrix(x, x_nm = NULL)

assert_is_number_nonNA_atom(x, x_nm = NULL)

assert_is_number_nonNA_gtezero_atom(x, x_nm = NULL)

assert_is_number_nonNA_gtezero_matrix(x, x_nm = NULL)

assert_is_number_nonNA_gtezero_vector(x, x_nm = NULL)

assert_is_number_nonNA_gtzero_atom(x, x_nm = NULL)

assert_is_number_nonNA_gtzero_matrix(x, x_nm = NULL)

assert_is_number_nonNA_gtzero_vector(x, x_nm = NULL)

assert_is_number_nonNA_ltezero_atom(x, x_nm = NULL)

assert_is_number_nonNA_ltezero_matrix(x, x_nm = NULL)

assert_is_number_nonNA_ltezero_vector(x, x_nm = NULL)

assert_is_number_nonNA_ltzero_atom(x, x_nm = NULL)

assert_is_number_nonNA_ltzero_matrix(x, x_nm = NULL)

assert_is_number_nonNA_ltzero_vector(x, x_nm = NULL)

assert_is_number_nonNA_matrix(x, x_nm = NULL)

assert_is_number_nonNA_vector(x, x_nm = NULL)

assert_is_number_vector(x, x_nm = NULL)

assert_is_numeric(x, x_nm = NULL)

assert_is_of_length(x, x_nm = NULL, expected_length)

assert_is_uniquely_named(x, x_nm = NULL)

assert_is_uniquely_named_list(x, x_nm = NULL)

assert_is_vector(x, x_nm = NULL)

assert_vector_elems_are_in_set(x, x_nm = NULL, set)

report_atom_is_in_set(x, x_nm = NULL, set)

report_dir_exists(x, x_nm = NULL)

report_file_exists(x, x_nm = NULL)

report_has_class(x, x_nm = NULL, required_class)

report_has_length(x, x_nm = NULL, expected_length)

report_has_names(x, x_nm = NULL, required_names)

report_has_one_of_classes(x, x_nm = NULL, classes)

report_has_only_names(x, x_nm = NULL, required_names)

report_inherits(x, x_nm = NULL, required_class)

report_is_atom(x, x_nm = NULL)

report_is_between_exclusive(x, x_nm = NULL, lo, hi)

report_is_between_inclusive(x, x_nm = NULL, lo, hi)

report_is_call(x, x_nm = NULL)

report_is_character(x, x_nm = NULL)

report_is_data.frame(x, x_nm = NULL)

report_is_data.frame_with_required_names(x, x_nm = NULL, required_names)

report_is_data.table(x, x_nm = NULL)

report_is_data.table_with_required_names(x, x_nm = NULL, required_names)

report_is_data_table(x, x_nm = NULL)

report_is_data_table_with_required_names(x, x_nm = NULL, required_names)

report_is_Date(x, x_nm = NULL)

report_is_double(x, x_nm = NULL)

report_is_environment(x, x_nm = NULL)

report_is_expression(x, x_nm = NULL)

report_is_factor(x, x_nm = NULL)

report_is_factor_with_levels(x, x_nm = NULL, expected_levels)

report_is_function(x, x_nm = NULL)

report_is_gt(x, x_nm = NULL, lo)

report_is_gte(x, x_nm = NULL, lo)

report_is_gtezero(x, x_nm = NULL)

report_is_gtzero(x, x_nm = NULL)

report_is_integer(x, x_nm = NULL)

report_is_language_object(x, x_nm = NULL)

report_is_list(x, x_nm = NULL)

report_is_logical(x, x_nm = NULL)

report_is_lt(x, x_nm = NULL, lo, hi)

report_is_lte(x, x_nm = NULL, hi)

report_is_ltezero(x, x_nm = NULL)

report_is_ltzero(x, x_nm = NULL)

report_is_matrix(x, x_nm = NULL)

report_is_name(x, x_nm = NULL)

report_is_named(x, x_nm = NULL)

report_is_named_list(x, x_nm = NULL)

report_is_nonNA(x, x_nm = NULL)

report_is_NULL(x, x_nm = NULL)

report_is_number(x, x_nm = NULL)

report_is_numeric(x, x_nm = NULL)

report_is_of_length(x, x_nm = NULL, expected_length)

report_is_uniquely_named(x, x_nm = NULL)

report_is_uniquely_named_list(x, x_nm = NULL)

report_is_vector(x, x_nm = NULL)

report_vector_elems_are_in_set(x, x_nm = NULL, set)

report_is_Date_atom(x, x_nm = NULL)

report_is_Date_matrix(x, x_nm = NULL)

report_is_Date_vector(x, x_nm = NULL)

report_is_Date_nonNA_atom(x, x_nm = NULL)

report_is_Date_nonNA_matrix(x, x_nm = NULL)

report_is_Date_nonNA_vector(x, x_nm = NULL)

report_is_character_atom(x, x_nm = NULL)

report_is_character_matrix(x, x_nm = NULL)

report_is_character_vector(x, x_nm = NULL)

report_is_character_nonNA_atom(x, x_nm = NULL)

report_is_character_nonNA_matrix(x, x_nm = NULL)

report_is_character_nonNA_vector(x, x_nm = NULL)

report_is_double_atom(x, x_nm = NULL)

report_is_double_matrix(x, x_nm = NULL)

report_is_double_vector(x, x_nm = NULL)

report_is_double_gtezero_atom(x, x_nm = NULL)

report_is_double_gtezero_matrix(x, x_nm = NULL)

report_is_double_gtezero_vector(x, x_nm = NULL)

report_is_double_gtzero_atom(x, x_nm = NULL)

report_is_double_gtzero_matrix(x, x_nm = NULL)

report_is_double_gtzero_vector(x, x_nm = NULL)

report_is_double_ltezero_atom(x, x_nm = NULL)

report_is_double_ltezero_matrix(x, x_nm = NULL)

report_is_double_ltezero_vector(x, x_nm = NULL)

report_is_double_ltzero_atom(x, x_nm = NULL)

report_is_double_ltzero_matrix(x, x_nm = NULL)

report_is_double_ltzero_vector(x, x_nm = NULL)

report_is_double_nonNA_atom(x, x_nm = NULL)

report_is_double_nonNA_matrix(x, x_nm = NULL)

report_is_double_nonNA_vector(x, x_nm = NULL)

report_is_double_nonNA_gtezero_atom(x, x_nm = NULL)

report_is_double_nonNA_gtezero_matrix(x, x_nm = NULL)

report_is_double_nonNA_gtezero_vector(x, x_nm = NULL)

report_is_double_nonNA_gtzero_atom(x, x_nm = NULL)

report_is_double_nonNA_gtzero_matrix(x, x_nm = NULL)

report_is_double_nonNA_gtzero_vector(x, x_nm = NULL)

report_is_double_nonNA_ltezero_atom(x, x_nm = NULL)

report_is_double_nonNA_ltezero_matrix(x, x_nm = NULL)

report_is_double_nonNA_ltezero_vector(x, x_nm = NULL)

report_is_double_nonNA_ltzero_atom(x, x_nm = NULL)

report_is_double_nonNA_ltzero_matrix(x, x_nm = NULL)

report_is_double_nonNA_ltzero_vector(x, x_nm = NULL)

report_is_factor_atom(x, x_nm = NULL)

report_is_factor_matrix(x, x_nm = NULL)

report_is_factor_vector(x, x_nm = NULL)

report_is_factor_nonNA_atom(x, x_nm = NULL)

report_is_factor_nonNA_matrix(x, x_nm = NULL)

report_is_factor_nonNA_vector(x, x_nm = NULL)

report_is_integer_atom(x, x_nm = NULL)

report_is_integer_matrix(x, x_nm = NULL)

report_is_integer_vector(x, x_nm = NULL)

report_is_integer_gtezero_atom(x, x_nm = NULL)

report_is_integer_gtezero_matrix(x, x_nm = NULL)

report_is_integer_gtezero_vector(x, x_nm = NULL)

report_is_integer_gtzero_atom(x, x_nm = NULL)

report_is_integer_gtzero_matrix(x, x_nm = NULL)

report_is_integer_gtzero_vector(x, x_nm = NULL)

report_is_integer_ltezero_atom(x, x_nm = NULL)

report_is_integer_ltezero_matrix(x, x_nm = NULL)

report_is_integer_ltezero_vector(x, x_nm = NULL)

report_is_integer_ltzero_atom(x, x_nm = NULL)

report_is_integer_ltzero_matrix(x, x_nm = NULL)

report_is_integer_ltzero_vector(x, x_nm = NULL)

report_is_integer_nonNA_atom(x, x_nm = NULL)

report_is_integer_nonNA_matrix(x, x_nm = NULL)

report_is_integer_nonNA_vector(x, x_nm = NULL)

report_is_integer_nonNA_gtezero_atom(x, x_nm = NULL)

report_is_integer_nonNA_gtezero_matrix(x, x_nm = NULL)

report_is_integer_nonNA_gtezero_vector(x, x_nm = NULL)

report_is_integer_nonNA_gtzero_atom(x, x_nm = NULL)

report_is_integer_nonNA_gtzero_matrix(x, x_nm = NULL)

report_is_integer_nonNA_gtzero_vector(x, x_nm = NULL)

report_is_integer_nonNA_ltezero_atom(x, x_nm = NULL)

report_is_integer_nonNA_ltezero_matrix(x, x_nm = NULL)

report_is_integer_nonNA_ltezero_vector(x, x_nm = NULL)

report_is_integer_nonNA_ltzero_atom(x, x_nm = NULL)

report_is_integer_nonNA_ltzero_matrix(x, x_nm = NULL)

report_is_integer_nonNA_ltzero_vector(x, x_nm = NULL)

report_is_logical_atom(x, x_nm = NULL)

report_is_logical_matrix(x, x_nm = NULL)

report_is_logical_vector(x, x_nm = NULL)

report_is_logical_nonNA_atom(x, x_nm = NULL)

report_is_logical_nonNA_matrix(x, x_nm = NULL)

report_is_logical_nonNA_vector(x, x_nm = NULL)

report_is_number_atom(x, x_nm = NULL)

report_is_number_matrix(x, x_nm = NULL)

report_is_number_vector(x, x_nm = NULL)

report_is_number_gtezero_atom(x, x_nm = NULL)

report_is_number_gtezero_matrix(x, x_nm = NULL)

report_is_number_gtezero_vector(x, x_nm = NULL)

report_is_number_gtzero_atom(x, x_nm = NULL)

report_is_number_gtzero_matrix(x, x_nm = NULL)

report_is_number_gtzero_vector(x, x_nm = NULL)

report_is_number_ltezero_atom(x, x_nm = NULL)

report_is_number_ltezero_matrix(x, x_nm = NULL)

report_is_number_ltezero_vector(x, x_nm = NULL)

report_is_number_ltzero_atom(x, x_nm = NULL)

report_is_number_ltzero_matrix(x, x_nm = NULL)

report_is_number_ltzero_vector(x, x_nm = NULL)

report_is_number_nonNA_atom(x, x_nm = NULL)

report_is_number_nonNA_matrix(x, x_nm = NULL)

report_is_number_nonNA_vector(x, x_nm = NULL)

report_is_number_nonNA_gtezero_atom(x, x_nm = NULL)

report_is_number_nonNA_gtezero_matrix(x, x_nm = NULL)

report_is_number_nonNA_gtezero_vector(x, x_nm = NULL)

report_is_number_nonNA_gtzero_atom(x, x_nm = NULL)

report_is_number_nonNA_gtzero_matrix(x, x_nm = NULL)

report_is_number_nonNA_gtzero_vector(x, x_nm = NULL)

report_is_number_nonNA_ltezero_atom(x, x_nm = NULL)

report_is_number_nonNA_ltezero_matrix(x, x_nm = NULL)

report_is_number_nonNA_ltezero_vector(x, x_nm = NULL)

report_is_number_nonNA_ltzero_atom(x, x_nm = NULL)

report_is_number_nonNA_ltzero_matrix(x, x_nm = NULL)

report_is_number_nonNA_ltzero_vector(x, x_nm = NULL)

test_atom_is_in_set(x, x_nm = NULL, set)

test_dir_exists(x, x_nm = NULL)

test_file_exists(x, x_nm = NULL)

test_has_class(x, x_nm = NULL, required_class)

test_has_length(x, x_nm = NULL, expected_length)

test_has_names(x, x_nm = NULL, required_names)

test_has_one_of_classes(x, x_nm = NULL, classes)

test_has_only_names(x, x_nm = NULL, required_names)

test_inherits(x, x_nm = NULL, required_class)

test_is_atom(x, x_nm = NULL)

test_is_between_exclusive(x, x_nm = NULL, lo, hi)

test_is_between_inclusive(x, x_nm = NULL, lo, hi)

test_is_call(x, x_nm = NULL)

test_is_character(x, x_nm = NULL)

test_is_character_atom(x, x_nm = NULL)

test_is_character_matrix(x, x_nm = NULL)

test_is_character_nonNA_atom(x, x_nm = NULL)

test_is_character_nonNA_matrix(x, x_nm = NULL)

test_is_character_nonNA_vector(x, x_nm = NULL)

test_is_character_vector(x, x_nm = NULL)

test_is_data.frame(x, x_nm = NULL)

test_is_data.frame_with_required_names(x, x_nm = NULL, required_names)

test_is_data.table(x, x_nm = NULL)

test_is_data.table_with_required_names(x, x_nm = NULL, required_names)

test_is_data_table(x, x_nm = NULL)

test_is_data_table_with_required_names(x, x_nm = NULL, required_names)

test_is_Date(x, x_nm = NULL)

test_is_Date_atom(x, x_nm = NULL)

test_is_Date_matrix(x, x_nm = NULL)

test_is_Date_nonNA_atom(x, x_nm = NULL)

test_is_Date_nonNA_matrix(x, x_nm = NULL)

test_is_Date_nonNA_vector(x, x_nm = NULL)

test_is_Date_vector(x, x_nm = NULL)

test_is_double(x, x_nm = NULL)

test_is_double_atom(x, x_nm = NULL)

test_is_double_gtezero_atom(x, x_nm = NULL)

test_is_double_gtezero_matrix(x, x_nm = NULL)

test_is_double_gtezero_vector(x, x_nm = NULL)

test_is_double_gtzero_atom(x, x_nm = NULL)

test_is_double_gtzero_matrix(x, x_nm = NULL)

test_is_double_gtzero_vector(x, x_nm = NULL)

test_is_double_ltezero_atom(x, x_nm = NULL)

test_is_double_ltezero_matrix(x, x_nm = NULL)

test_is_double_ltezero_vector(x, x_nm = NULL)

test_is_double_ltzero_atom(x, x_nm = NULL)

test_is_double_ltzero_matrix(x, x_nm = NULL)

test_is_double_ltzero_vector(x, x_nm = NULL)

test_is_double_matrix(x, x_nm = NULL)

test_is_double_nonNA_atom(x, x_nm = NULL)

test_is_double_nonNA_gtezero_atom(x, x_nm = NULL)

test_is_double_nonNA_gtezero_matrix(x, x_nm = NULL)

test_is_double_nonNA_gtezero_vector(x, x_nm = NULL)

test_is_double_nonNA_gtzero_atom(x, x_nm = NULL)

test_is_double_nonNA_gtzero_matrix(x, x_nm = NULL)

test_is_double_nonNA_gtzero_vector(x, x_nm = NULL)

test_is_double_nonNA_ltezero_atom(x, x_nm = NULL)

test_is_double_nonNA_ltezero_matrix(x, x_nm = NULL)

test_is_double_nonNA_ltezero_vector(x, x_nm = NULL)

test_is_double_nonNA_ltzero_atom(x, x_nm = NULL)

test_is_double_nonNA_ltzero_matrix(x, x_nm = NULL)

test_is_double_nonNA_ltzero_vector(x, x_nm = NULL)

test_is_double_nonNA_matrix(x, x_nm = NULL)

test_is_double_nonNA_vector(x, x_nm = NULL)

test_is_double_vector(x, x_nm = NULL)

test_is_environment(x, x_nm = NULL)

test_is_expression(x, x_nm = NULL)

test_is_factor(x, x_nm = NULL)

test_is_factor_atom(x, x_nm = NULL)

test_is_factor_matrix(x, x_nm = NULL)

test_is_factor_nonNA_atom(x, x_nm = NULL)

test_is_factor_nonNA_matrix(x, x_nm = NULL)

test_is_factor_nonNA_vector(x, x_nm = NULL)

test_is_factor_vector(x, x_nm = NULL)

test_is_factor_with_levels(x, x_nm = NULL, expected_levels)

test_is_function(x, x_nm = NULL)

test_is_gt(x, x_nm = NULL, lo)

test_is_gte(x, x_nm = NULL, lo)

test_is_gtezero(x, x_nm = NULL)

test_is_gtzero(x, x_nm = NULL)

test_is_integer(x, x_nm = NULL)

test_is_integer_atom(x, x_nm = NULL)

test_is_integer_gtezero_atom(x, x_nm = NULL)

test_is_integer_gtezero_matrix(x, x_nm = NULL)

test_is_integer_gtezero_vector(x, x_nm = NULL)

test_is_integer_gtzero_atom(x, x_nm = NULL)

test_is_integer_gtzero_matrix(x, x_nm = NULL)

test_is_integer_gtzero_vector(x, x_nm = NULL)

test_is_integer_ltezero_atom(x, x_nm = NULL)

test_is_integer_ltezero_matrix(x, x_nm = NULL)

test_is_integer_ltezero_vector(x, x_nm = NULL)

test_is_integer_ltzero_atom(x, x_nm = NULL)

test_is_integer_ltzero_matrix(x, x_nm = NULL)

test_is_integer_ltzero_vector(x, x_nm = NULL)

test_is_integer_matrix(x, x_nm = NULL)

test_is_integer_nonNA_atom(x, x_nm = NULL)

test_is_integer_nonNA_gtezero_atom(x, x_nm = NULL)

test_is_integer_nonNA_gtezero_matrix(x, x_nm = NULL)

test_is_integer_nonNA_gtezero_vector(x, x_nm = NULL)

test_is_integer_nonNA_gtzero_atom(x, x_nm = NULL)

test_is_integer_nonNA_gtzero_matrix(x, x_nm = NULL)

test_is_integer_nonNA_gtzero_vector(x, x_nm = NULL)

test_is_integer_nonNA_ltezero_atom(x, x_nm = NULL)

test_is_integer_nonNA_ltezero_matrix(x, x_nm = NULL)

test_is_integer_nonNA_ltezero_vector(x, x_nm = NULL)

test_is_integer_nonNA_ltzero_atom(x, x_nm = NULL)

test_is_integer_nonNA_ltzero_matrix(x, x_nm = NULL)

test_is_integer_nonNA_ltzero_vector(x, x_nm = NULL)

test_is_integer_nonNA_matrix(x, x_nm = NULL)

test_is_integer_nonNA_vector(x, x_nm = NULL)

test_is_integer_vector(x, x_nm = NULL)

test_is_language_object(x, x_nm = NULL)

test_is_list(x, x_nm = NULL)

test_is_logical(x, x_nm = NULL)

test_is_logical_atom(x, x_nm = NULL)

test_is_logical_matrix(x, x_nm = NULL)

test_is_logical_nonNA_atom(x, x_nm = NULL)

test_is_logical_nonNA_matrix(x, x_nm = NULL)

test_is_logical_nonNA_vector(x, x_nm = NULL)

test_is_logical_vector(x, x_nm = NULL)

test_is_lt(x, x_nm = NULL, lo, hi)

test_is_lte(x, x_nm = NULL, hi)

test_is_ltezero(x, x_nm = NULL)

test_is_ltzero(x, x_nm = NULL)

test_is_matrix(x, x_nm = NULL)

test_is_name(x, x_nm = NULL)

test_is_named(x, x_nm = NULL)

test_is_named_list(x, x_nm = NULL)

test_is_nonNA(x, x_nm = NULL)

test_is_NULL(x, x_nm = NULL)

test_is_number(x, x_nm = NULL)

test_is_number_atom(x, x_nm = NULL)

test_is_number_gtezero_atom(x, x_nm = NULL)

test_is_number_gtezero_matrix(x, x_nm = NULL)

test_is_number_gtezero_vector(x, x_nm = NULL)

test_is_number_gtzero_atom(x, x_nm = NULL)

test_is_number_gtzero_matrix(x, x_nm = NULL)

test_is_number_gtzero_vector(x, x_nm = NULL)

test_is_number_ltezero_atom(x, x_nm = NULL)

test_is_number_ltezero_matrix(x, x_nm = NULL)

test_is_number_ltezero_vector(x, x_nm = NULL)

test_is_number_ltzero_atom(x, x_nm = NULL)

test_is_number_ltzero_matrix(x, x_nm = NULL)

test_is_number_ltzero_vector(x, x_nm = NULL)

test_is_number_matrix(x, x_nm = NULL)

test_is_number_nonNA_atom(x, x_nm = NULL)

test_is_number_nonNA_gtezero_atom(x, x_nm = NULL)

test_is_number_nonNA_gtezero_matrix(x, x_nm = NULL)

test_is_number_nonNA_gtezero_vector(x, x_nm = NULL)

test_is_number_nonNA_gtzero_atom(x, x_nm = NULL)

test_is_number_nonNA_gtzero_matrix(x, x_nm = NULL)

test_is_number_nonNA_gtzero_vector(x, x_nm = NULL)

test_is_number_nonNA_ltezero_atom(x, x_nm = NULL)

test_is_number_nonNA_ltezero_matrix(x, x_nm = NULL)

test_is_number_nonNA_ltezero_vector(x, x_nm = NULL)

test_is_number_nonNA_ltzero_atom(x, x_nm = NULL)

test_is_number_nonNA_ltzero_matrix(x, x_nm = NULL)

test_is_number_nonNA_ltzero_vector(x, x_nm = NULL)

test_is_number_nonNA_matrix(x, x_nm = NULL)

test_is_number_nonNA_vector(x, x_nm = NULL)

test_is_number_vector(x, x_nm = NULL)

test_is_numeric(x, x_nm = NULL)

test_is_of_length(x, x_nm = NULL, expected_length)

test_is_uniquely_named(x, x_nm = NULL)

test_is_uniquely_named_list(x, x_nm = NULL)

test_is_vector(x, x_nm = NULL)

test_vector_elems_are_in_set(x, x_nm = NULL, set)

assert_is_one_of(x, x_nm = NULL, fun_nms, funs_arg_list = NULL)
}
\arguments{
\item{x}{R object to be tested (mandatory, no default)}

\item{x_nm}{\verb{[NULL, character]} (optional, default \code{NULL})

the name of the object \code{x} to mention in possible error message.
\itemize{
\item \code{NULL}: taken as \code{deparse(substitute(x))}
\item \code{character}: the name as a string
}}

\item{set}{\verb{[any vector]} (mandatory, no default)
set of values to compare to}

\item{required_class}{\verb{[character]} (mandatory, no default)
class that object must have}

\item{expected_length}{\verb{[integer]} (mandatory, no default)
length object must have}

\item{required_names}{\verb{[character]} (mandatory, no default)
set of names object must have}

\item{classes}{\verb{[character]} (mandatory, no default)
one or more classes; object must have at least one of these as class}

\item{lo}{\verb{[number]} (mandatory, no default)
lower bound for \code{x}}

\item{hi}{\verb{[number]} (mandatory, no default)
upper bound for \code{x}}

\item{expected_levels}{\verb{[character]} (mandatory, no default)
set of levels factor is required to have}

\item{fun_nms}{\verb{[character]} (mandatory, no default)

names of assertion funs to run}

\item{funs_arg_list}{\verb{[list, NULL]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no additional arguments are passed to functions specified in
\code{fun_nms}
\item \code{list}: named elements of this list are matched to the names of arguments
of each function specified in \code{fun_nms}, and matching arguments are
used; e.g. if one fun has arg \code{y} and other has \code{z}, you may pass
\code{list(y = my_y, z = my_z)}.
}}
}
\description{
Collection of assertion functions which raise an error if the test does not
pass.
}
