% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manually_written_dataset_checks.R
\name{dataset_validation}
\alias{dataset_validation}
\alias{identify_invalid_observations}
\title{Dataset Validation}
\usage{
identify_invalid_observations(x, report_df)
}
\arguments{
\item{x}{\verb{[data.frame]} (mandatory, no default)
dataset to validate}

\item{report_df}{\verb{[data.frame]} (mandatory, no default)

a report \code{data.frame} as produced by e.g. \link{tests_to_report}}
}
\value{
\itemize{
\item \code{identify_invalid_observations}: returns a data.frame with columns
\code{is_valid}, \code{fail_test_set}, and \code{fail_message_set}; new columns are
populated based on \code{report_df} contents; the data.frame has \code{nrow(x)} rows
}
}
\description{
Functions to ensure that a dataset (data.frame) looks as it should.
}
