% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{generate_assertions}
\alias{generate_assertions}
\title{Generate Assertions}
\usage{
generate_assertions(
  source_script = "R/assertions.R",
  target_script = "R/generated_assertions.R",
  pad = rep("", 5)
)
}
\arguments{
\item{source_script}{\verb{[character]} (mandatory, no default)

path to R script where your assertion functions are}

\item{target_script}{\verb{[character]} (mandatory, default \code{"assertions.R"})

path where generated code will be written}

\item{pad}{\verb{[character]} (mandatory, default \code{rep("", 5)})

lines between generated function definitions; by default 5 empty lines}
}
\description{
Generate code for a bunch of assertions functions.
}
