% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generation.R
\name{generate_function_variants}
\alias{generate_function_variants}
\title{Generate Assertions}
\usage{
generate_function_variants(
  prefix = c("report", "assert", "test")[1],
  target_script = "R/generated_report_fun_variants.R",
  pad = rep("", 5)
)
}
\arguments{
\item{target_script}{\verb{[character]} (mandatory, default \code{"assertions.R"})

path where generated code will be written}

\item{pad}{\verb{[character]} (mandatory, default \code{rep("", 5)})

lines between generated function definitions; by default 5 empty lines}

\item{source_script}{\verb{[character]} (mandatory, no default)

path to R script where your assertion functions are}
}
\description{
Generate code for a bunch of assertions functions.
}
