% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manually_written_check_funs.R
\name{is_between}
\alias{is_between}
\alias{is_between_inclusive}
\alias{is_between_exclusive}
\title{Between}
\usage{
is_between(x, lo, hi, inclusive = TRUE)

is_between_inclusive(x, lo, hi)

is_between_exclusive(x, lo, hi)
}
\arguments{
\item{x}{\verb{[R object]} (mandatory, no default)
e.g. numeric vector; each element is tested for being between \code{lo} and \code{hi}}

\item{lo}{\verb{[R object]} (mandatory, no default)
lower bound of length 1 or \code{length(x)}}

\item{hi}{\verb{[R object]} (mandatory, no default)
upper bound of length 1 or \code{length(x)}}

\item{inclusive}{\verb{[logical]} (optional, default \code{TRUE})
if \code{TRUE}, test is \code{lo <= x & x <= hi} and else \code{lo < x & x < hi}.}
}
\description{
Test if elements are between upper and lower bounds.
}
\section{Functions}{
\itemize{
\item \code{is_between_inclusive}: \code{is_between} with \code{inclusive = TRUE}

\item \code{is_between_exclusive}: \code{is_between} with \code{inclusive = FALSE}
}}

